Refactor: 중복되는 useLoginState 훅 제거 및 훅 네이밍 변경#51
